Moved auth related classes and interfaces to the new package
