+++ src/main/java/com/battlemetrics/authorization/JwtService.java
@@ -1,4 +1,4 @@
-package com.battlemetrics.service;
+package com.battlemetrics.authorization;
 
 import org.springframework.security.core.userdetails.UserDetails;
