+++ src/main/java/com/battlemetrics/config/SecurityConfiguration.java
@@ -2,7 +2,7 @@ package com.battlemetrics.config;
 
 import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
-import com.battlemetrics.service.UserService;
+import com.battlemetrics.authorization.UserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
